requirements.txt
boto3
numpy
torch
torchvision
torchaudio
Pillow
imutils
opencv-python
facenet-pytorch